Support for gen-mg for multi-asic linecards
For multi-asic linecards the following fields in the inventory (from TestbedProcessing.py) are converted to lists instead of a single value:
  - switchids - instead of start_switchid
  - voq_inband_ip
  - voq_inband_ipv6
  - voq_inband_intf

We also added the following fields:
  - loopback4096_ip
  - loopback4096_ipv6

The minigraph templates were modified to use generate asic level metadata for all asics in a linecard.
We added minigraph_dpg_voq_asic.j2 template to generate asic DPG definitions for multi-asic linecards.